Update README.md
I made tic tac toe game using python language. I used simply the basic concepts of python data types, if-else, for or while statements and function. The program code is very simply and easy to understandable. And this project is very helpful to you for understanding the basic concepts of python.